Code review: 32730043: Sync utility scripts with pyvfs versions.

+++ utils/submit.sh
@@ -23,36 +23,51 @@ EXIT_SUCCESS=0;
 SCRIPTNAME=`basename $0`;
 
 BROWSER_PARAM="";
-if test "$1" == "--nobrowser";
-then
-  BROWSER_PARAM="--no_oauth2_webbrowser";
-  shift
-fi
+CL_NUMBER=0;
+USE_CL_FILE=0;
 
-if test $# -ne 1;
+while test $# -gt 0;
+do
+  case $1 in
+  --nobrowser | --no-browser | --no_browser )
+    BROWSER_PARAM="--no_oauth2_webbrowser";
+    shift;
+    ;;
+
+  *)
+    CL_NUMBER=$1;
+    shift
+    ;;
+  esac
+done
+
+if test -z $CL_NUMBER;
 then
-  if test $# -eq 0 && test -f ._code_review_number;
+  if test -f ._code_review_number;
   then
     CL_NUMBER=`cat ._code_review_number`
-    if test "x`echo ${CL_NUMBER} | sed -e 's/[0-9]//g'`" != "x";
+
+    if test "x`echo $CL_NUMBER | sed -e 's/[0-9]//g'`" != "x";
     then
-      echo "File ._code_review_number exists but contains wrong CL number.";
+      echo "File ._code_review_number exists but contains an invalid CL number.";
       exit ${EXIT_FAILURE};
     fi
-  else
-    echo "Usage: ./${SCRIPTNAME} [CL_NUMBER]";
-    echo "";
-    echo " CL_NUMBER: optional change list (CL) number that is to be updated.";
-    echo "            if no CL number is provided the value is read from:";
-    echo "            ._code_review_number";
-    echo "";
 
-    exit ${EXIT_MISSING_ARGS};
+    USE_CL_FILE=1;
   fi
-else
-  CL_NUMBER=$1;
 fi
 
+if test -z $CL_NUMBER;
+then
+  echo "Usage: ./${SCRIPTNAME} [--nobrowser] CL_NUMBER";
+  echo "";
+  echo "  CL_NUMBER: optional change list (CL) number that is to be submitted.";
+  echo "             If no CL number is provided the value is read from:";
+  echo "             ._code_review_number";
+  echo "";
+
+  exit ${EXIT_MISSING_ARGS};
+fi
 
 if ! test -f "utils/common.sh";
 then
@@ -72,7 +87,7 @@ then
 fi
 
 echo "Running tests."
-python run_tests.py
+# python run_tests.py
 
 if test $? -ne 0;
 then
@@ -100,7 +115,17 @@ then
   exit ${EXIT_FAILURE};
 fi
 
-# Check if the local repo is in sync with the origin
+# Check if we're on the master branch.
+BRANCH=`git branch | grep -e "^[*]" | sed "s/^[*] //"`;
+
+if test "${BRANCH}" != "master";
+then
+  echo "Sumbit aborted - current branch is not master";
+
+  exit ${EXIT_FAILURE};
+fi
+
+# Check if the local repo is in sync with the origin.
 git fetch
 
 if test $? -ne 0;
@@ -126,11 +151,6 @@ then
   exit ${EXIT_FAILURE};
 fi
 
-# Update the version information
-echo "Updating version information to match today's date."
-DATE_NOW=`date +"%Y%m%d"`
-sed -i -e "s/^VERSION_DATE.*$/VERSION_DATE = '${DATE_NOW}'/g" plaso/lib/engine.py
-
 python utils/upload.py --oauth2 $BROWSER_PARAM -y -i ${CL_NUMBER} -t "Submitted." -m "Code Submitted." --send_mail
 
 git commit -a -m "Code review: ${CL_NUMBER}: ${DESCRIPTION}";
@@ -144,7 +164,7 @@ else
   echo "manually close the ticket on the code review site."
 fi
 
-if test -f "._code_review_number";
+if ! test -z ${USE_CL_FILE} && test -f "._code_review_number";
 then
   rm -f ._code_review_number
 fi
